requirements.txt
click==8.0.4
finch==0.3.3
matplotlib==3.5.2
numpy==1.21.5
Pillow==9.3.0
scikit_learn==1.2.0
torch==1.13.1
torchvision==0.14.1
